Use SES Signature Version 4 via boto3
AWS has deprecated SES Signature Version 4 for API requests, and it will stop working September 30, 2020.

Boto2 seems to not support v4 (https://stackoverflow.com/a/63454370/729729), so we need to switch to boto3. The two boto versions can coexist, so this isn't a complete switch from one to the other -- only where necessary.